move node class to separate file
